requirements.txt
Django==5.2.6
celery[librabbitmq]==5.5.3
redis==5.2.1
prometheus-client==0.22.1
django-prometheus==2.4.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
Pillow==11.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
stripe==7.9.0
